# packages/shared/openapi.yaml
openapi: 3.0.3
info:
  title: Meme Manager API
  description: 表情包管理工具云端同步 API
  version: 1.0.0
  contact:
    name: Meme Manager Team
  license:
    name: MIT
servers:
  - url: https://api.meme-manager.example.com
    description: Production server
  - url: https://staging-api.meme-manager.example.com
    description: Staging server
  - url: http://localhost:8787
    description: Development server
paths:
...
  /index:
    get:
      summary: 拉取事件
      description: 获取指定时钟之后的事件列表
      operationId: getEvents
      tags:
        - Sync
      security:
        - BearerAuth: []
      parameters:
        - name: since
          in: query
          description: 起始服务器时钟
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: 返回事件数量限制
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: deviceId
          in: query
          description: 过滤特定设备的事件
          required: false
          schema:
            type: string
            format: uuid
        - name: types
          in: query
          description: 过滤事件类型
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: 事件列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IndexResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

    post:
      summary: 推送事件
      description: 批量提交本地事件到服务器
      operationId: pushEvents
      tags:
        - Sync
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexBatchRequest'
      responses:
        '200':
          description: 事件提交结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IndexBatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 同步冲突
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        allOf:
                          - $ref: '#/components/schemas/ApiError'
                          - type: object
                            properties:
                              code:
                                enum: [SYNC_CONFLICT]
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 设备访问令牌
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    allOf:
                      - $ref: '#/components/schemas/ApiError'
                      - type: object
                        properties:
                          code:
                            enum: [INVALID_REQUEST, VALIDATION_ERROR]
    Unauthorized:
      description: 未授权访问
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    allOf:
                      - $ref: '#/components/schemas/ApiError'
                      - type: object
                        properties:
                          code:
                            enum: [UNAUTHORIZED, INVALID_TOKEN, TOKEN_EXPIRED]
    RateLimited:
      description: 请求频率限制
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    allOf:
                      - $ref: '#/components/schemas/ApiError'
                      - type: object
                        properties:
                          code:
                            enum: [RATE_LIMITED]
      headers:
        Retry-After:
          description: 重试等待时间（秒）
          schema:
            type: integer
  schemas:
    ApiResponse:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          description: 请求是否成功
        data:
          description: 响应数据
        error:
          $ref: '#/components/schemas/ApiError'
        timestamp:
          type: number
          description: 响应时间戳
        requestId:
          type: string
          format: uuid
          description: 请求 ID
    ApiError:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          description: 错误码
        message:
          type: string
          description: 错误消息
        details:
          description: 错误详情
        timestamp:
          type: number
          description: 错误时间戳
        requestId:
          type: string
          format: uuid
          description: 请求 ID
...
    IndexResponse:
      type: object
      required:
        - events
        - nextClock
        - hasMore
        - serverTime
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/OperationEvent'
          description: 事件列表
        nextClock:
          type: integer
          minimum: 0
          description: 下一个时钟值
        hasMore:
          type: boolean
          description: 是否还有更多事件
        serverTime:
          type: number
          description: 服务器当前时间
    IndexBatchRequest:
      type: object
      required:
        - events
        - clientClock
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/OperationEvent'
          minItems: 1
          maxItems: 100
          description: 事件列表
        clientClock:
          type: integer
          minimum: 0
          description: 客户端当前时钟
        expectServerClock:
          type: integer
          minimum: 0
          description: 期望的服务器时钟
    IndexBatchResponse:
      type: object
      required:
        - mappings
        - serverClock
      properties:
        mappings:
          type: array
          items:
            $ref: '#/components/schemas/EventMapping'
          description: 事件映射结果
        serverClock:
          type: integer
          minimum: 0
          description: 当前服务器时钟
        conflicts:
          type: array
          items:
            type: object
            required:
              - eventId
              - reason
            properties:
              eventId:
                type: string
                format: uuid
              reason:
                type: string
              suggestedResolution:
                description: 建议的解决方案
          description: 冲突列表
    EventMapping:
      type: object
      required:
        - clientEventId
        - serverEventId
        - serverClock
        - status
      properties:
        clientEventId:
          type: string
          format: uuid
          description: 客户端事件 ID
        serverEventId:
          type: string
          format: uuid
          description: 服务器事件 ID
        serverClock:
          type: integer
          minimum: 0
          description: 服务器时钟
        status:
          type: string
          enum: [accepted, duplicate, conflict, rejected]
          description: 处理状态
        conflictReason:
          type: string
          description: 冲突原因
    OperationEvent:
      type: object
      required:
        - id
        - type
        - operation
        - deviceId
        - timestamp
        - version
        - payload
      properties:
        id:
          type: string
          format: uuid
          description: 事件 ID
        type:
          type: string
          description: 事件类型
        operation:
          type: string
          enum: [create, update, delete, restore, add_relation, remove_relation, update_relation, reorder]
          description: 操作类型
        deviceId:
          type: string
          format: uuid
          description: 设备 ID
        userId:
          type: string
          format: uuid
          description: 用户 ID
        timestamp:
          type: number
          description: 事件时间戳
        serverTimestamp:
          type: number
          description: 服务器时间戳
        version:
          type: integer
          minimum: 1
          description: 事件版本
        causedBy:
          type: string
          format: uuid
          description: 引起此事件的事件 ID
        batchId:
          type: string
          format: uuid
          description: 批量操作 ID
        payload:
          description: 事件载荷
...